% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kana.R
\name{flash.hiragana}
\alias{flash.hiragana}
\title{flash.hiragana}
\usage{
flash.hiragana(size = 10, reduce = NULL, exclude = NULL)
}
\arguments{
\item{size}{TODO}

\item{reduce}{TODO}

\item{exclude}{TODO}
}
\description{
flash.hiragana
}

