% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kana.R
\name{flash.kana}
\alias{flash.kana}
\title{flash.hiragana}
\usage{
flash.kana(size = 10, romaji = NULL, kana = NULL, reduce = NULL,
  exclude = NULL)
}
\arguments{
\item{size}{TODO}

\item{reduce}{TODO}

\item{exclude}{TODO}
}
\description{
flash.hiragana
}

