% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kana.R
\name{flash.katakana}
\alias{flash.katakana}
\title{flash.hiragana}
\usage{
flash.katakana(size = 10, reduce = NULL, exclude = NULL)
}
\arguments{
\item{size}{TODO}

\item{reduce}{TODO}

\item{exclude}{TODO}
}
\description{
flash.hiragana
}

