% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanku.r
\docType{package}
\name{hanku}
\alias{hanku}
\alias{hanku-package}
\title{Japanese kana learning utilities}
\description{
hanku
}
\details{
Functions to assist with learning Japanese kana.
}
\author{
Christian Heckendorf \email{heckendorfc@gmail.com}
}
\keyword{package}

