% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kana.R
\name{kanasample}
\alias{kanasample}
\title{kanasample}
\usage{
kanasample(ro, kana, size, reduce, exclude)
}
\arguments{
\item{ro}{TODO}

\item{kana}{TODO}

\item{size}{TODO}

\item{reduce}{TODO}

\item{exclude}{TODO}
}
\description{
kanasample
}

