% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hanku.r
\docType{package}
\name{lineSampler}
\alias{lineSampler}
\alias{lineSampler-package}
\title{File Line Sampler}
\description{
hanku
}
\details{
Functions to assist with learning Japanese kana.
}
\author{
Drew Schmidt \email{wrathematics@gmail.com}
}
\keyword{package}

