% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kana.R
\name{trial.hiragana}
\alias{trial.hiragana}
\title{trial.hiragana}
\usage{
trial.hiragana(time = 30, reduce = NULL, exclude = NULL)
}
\arguments{
\item{time}{TODO}

\item{reduce}{TODO}

\item{exclude}{TODO}
}
\description{
trial.hiragana
}

