% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kana.R
\name{trial.kana}
\alias{trial.kana}
\title{trial.hiragana}
\usage{
trial.kana(time = 30, kana = NULL, romaji = NULL, reduce = NULL,
  exclude = NULL)
}
\arguments{
\item{time}{TODO}

\item{reduce}{TODO}

\item{exclude}{TODO}
}
\description{
trial.hiragana
}

