% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kana.R
\name{trial.katakana}
\alias{trial.katakana}
\title{trial.hiragana}
\usage{
trial.katakana(time = 30, reduce = NULL, exclude = NULL)
}
\arguments{
\item{time}{TODO}

\item{reduce}{TODO}

\item{exclude}{TODO}
}
\description{
trial.hiragana
}

